test: refactors tests for handle_sd
