+++ scripts/build_env/tests/sd/test_handle_sd_artifact.py
@@ -4,7 +4,7 @@ import pytest
 from unittest.mock import patch
 from ruamel.yaml import YAML
 
-from test_helper import TestHelpers
+from test_sd_helpers import do_prerequisites, assert_sd_contents, load_test_pipeline_sd_data
 
 os.environ['ENVIRONMENT_NAME'] = "temporary"
 os.environ['CLUSTER_NAME'] = "temporary"
@@ -41,12 +41,12 @@ SD = "sd.yaml"
 @patch("handle_sd.download_sd_by_appver")
 def test_sd_positive(mock_download_sd, test_case_name):
     env = Environment(str(Path(OUTPUT_DIR, test_case_name)), "cluster-01", "env-01")
-    TestHelpers.do_prerequisites(SD, TEST_SD_DIR, OUTPUT_DIR, test_case_name, env, test_suits_map)
+    do_prerequisites(SD, TEST_SD_DIR, OUTPUT_DIR, test_case_name, env, test_suits_map)
     logger.info(f"======TEST HANDLE_SD_ARTIFACT_POSITIVE: {test_case_name}======")
     logger.info(f"Starting SD test:"
                 f"\n\tTest case: {test_case_name}")
 
-    sd_data, sd_source_type, sd_version, sd_delta, sd_merge_mode = TestHelpers.load_test_pipeline_sd_data(TEST_SD_DIR, test_case_name)
+    sd_data, sd_source_type, sd_version, sd_delta, sd_merge_mode = load_test_pipeline_sd_data(TEST_SD_DIR, test_case_name)
 
     file_path = Path(TEST_SD_DIR, test_case_name, f"mock_sd.json")
     sd_data = openJson(file_path)
@@ -55,7 +55,7 @@ def test_sd_positive(mock_download_sd, test_case_name):
     handle_sd(env, sd_source_type, sd_version, sd_data, sd_delta, sd_merge_mode)
     actual_dir = os.path.join(env.env_path, "Inventory", "solution-descriptor")
 
-    TestHelpers.assert_sd_contents(TEST_SD_DIR, OUTPUT_DIR, test_case_name, actual_dir, test_suits_map)
+    assert_sd_contents(TEST_SD_DIR, OUTPUT_DIR, test_case_name, actual_dir, test_suits_map)
     logger.info(f"=====SUCCESS - {test_case_name}======")
     
     
@@ -63,12 +63,12 @@ def test_sd_positive(mock_download_sd, test_case_name):
 @patch("handle_sd.download_sd_by_appver")
 def test_sd_negative(mock_download_sd, test_case_name, expected_exception):
     env = Environment(str(Path(OUTPUT_DIR, test_case_name)), "cluster-01", "env-01")
-    TestHelpers.do_prerequisites(SD, TEST_SD_DIR, OUTPUT_DIR, test_case_name, env, test_suits_map)
+    do_prerequisites(SD, TEST_SD_DIR, OUTPUT_DIR, test_case_name, env, test_suits_map)
     logger.info(f"======TEST HANDLE_SD_ARTIFACT_NEGATIVE: {test_case_name}======")
     logger.info(f"Starting SD test:"
                 f"\n\tTest case: {test_case_name}")
 
-    sd_data, sd_source_type, sd_version, sd_delta, sd_merge_mode = TestHelpers.load_test_pipeline_sd_data(TEST_SD_DIR, test_case_name)
+    sd_data, sd_source_type, sd_version, sd_delta, sd_merge_mode = load_test_pipeline_sd_data(TEST_SD_DIR, test_case_name)
 
     file_path = Path(TEST_SD_DIR, test_case_name, f"mock_sd.json")
     sd_data = openJson(file_path)
