+++ scripts/build_env/tests/sd/test_handle_sd_local.py
@@ -3,7 +3,7 @@ import os
 import pytest
 from ruamel.yaml import YAML
 
-from test_helper import TestHelpers
+from test_sd_helpers import do_prerequisites, assert_sd_contents, load_test_pipeline_sd_data
 
 os.environ['ENVIRONMENT_NAME'] = "temporary"
 os.environ['CLUSTER_NAME'] = "temporary"
@@ -26,8 +26,6 @@ TEST_CASES_POSITIVE = [
     "TC-001-017",
 ]
 
-TEST_CASES_NEGATIVE = {}
-
 test_suits_map = {
     "basic_not_first": ["TC-001-010", "TC-001-012"],
     "basic_first": ["TC-001-002", "TC-001-004"],
@@ -44,35 +42,15 @@ SD = "sd.yaml"
 @pytest.mark.parametrize("test_case_name", TEST_CASES_POSITIVE)
 def test_sd_positive(test_case_name):
     env = Environment(str(Path(OUTPUT_DIR, test_case_name)), "cluster-01", "env-01")
-    TestHelpers.do_prerequisites(SD, TEST_SD_DIR, OUTPUT_DIR, test_case_name, env, test_suits_map)
+    do_prerequisites(SD, TEST_SD_DIR, OUTPUT_DIR, test_case_name, env, test_suits_map)
     logger.info(f"======TEST HANDLE_SD_LOCAL_POSITIVE: {test_case_name}======")
     logger.info(f"Starting SD test:"
                 f"\n\tTest case: {test_case_name}")
 
-    sd_data, sd_source_type, sd_version, sd_delta, sd_merge_mode = TestHelpers.load_test_pipeline_sd_data(TEST_SD_DIR, test_case_name)
+    sd_data, sd_source_type, sd_version, sd_delta, sd_merge_mode = load_test_pipeline_sd_data(TEST_SD_DIR, test_case_name)
         
     handle_sd(env, sd_source_type, sd_version, sd_data, sd_delta, sd_merge_mode)
     actual_dir = os.path.join(env.env_path, "Inventory", "solution-descriptor")
 
-    TestHelpers.assert_sd_contents(TEST_SD_DIR, OUTPUT_DIR, test_case_name, actual_dir, test_suits_map)
-    logger.info(f"=====SUCCESS - {test_case_name}======")
-    
-
-@pytest.mark.parametrize("test_case_name", TEST_CASES_NEGATIVE)
-def test_sd_negative(test_case_name):
-    env = Environment(str(Path(OUTPUT_DIR, test_case_name)), "cluster-01", "env-01")
-    TestHelpers.do_prerequisites(SD, TEST_SD_DIR, OUTPUT_DIR, test_case_name, env, test_suits_map)
-    logger.info(f"======TEST HANDLE_SD_LOCAL_NEGATIVE: {test_case_name}======")
-    logger.info(f"Starting SD test:"
-                f"\n\tTest case: {test_case_name}")
-
-    sd_data, sd_source_type, sd_version, sd_delta, sd_merge_mode = TestHelpers.load_test_pipeline_sd_data(TEST_SD_DIR, test_case_name)
-
-    file_path = Path(TEST_SD_DIR, test_case_name, f"mock_sd.json")
-    sd_data = openJson(file_path)
-    mock_download_sd.return_value = sd_data
-    
-    with pytest.raises(expected_exception):
-        handle_sd(env, sd_source_type, sd_version, sd_data, sd_delta, sd_merge_mode)
-
+    assert_sd_contents(TEST_SD_DIR, OUTPUT_DIR, test_case_name, actual_dir, test_suits_map)
     logger.info(f"=====SUCCESS - {test_case_name}======")
